requirements.txt
browser-use>=0.1.0
playwright>=1.40.0
python-dotenv>=1.0.0
